Entregas - Rogério Muniz/Projeto Final/requirements.txt
streamlit
pytesseract
pdfplumber
Pillow
requests
folium
streamlit-folium
shapely
pyproj
pandas
llama-cpp-python